+++ examples/storybook/src/shared/resources/experimentsExample.ts
@@ -20,4 +20,5 @@ export const EXPERIMENTS = [
   { name: 'useFilePluginAutoDownloadLinkRef', scope: 'ricos' },
   { name: 'enableFilePluginPDFViewer', scope: 'ricos' },
   { name: 'encAutoImageUrls', scope: 'ricos' },
+  { name: 'imageFigureCaption', scope: 'ricos' },
 ];