+++ CHANGELOG.md
@@ -17,7 +17,10 @@
     Click to see more.
   </summary>
 
+### :nail_care: Polish
 
+- `image`
+  - [#3740](https://github.com/wix/ricos/pull/3740) SEO optimization: render image caption as figure caption
 
 </details>
   <hr/>
@@ -25,6 +28,7 @@
 ## 8.70.41 (April 17, 2022)
 
 ### :bug: Bug Fix
+
 - `video`
   - [#3737](https://github.com/wix/ricos/pull/#3737) fix "getInternalPlayer().play(), play is not a function"
 - `audio`
@@ -276,7 +280,6 @@
   - [#3618](https://github.com/wix/ricos/pull/3618) add audio modal settings modal
 - `polls`
   - [#3622](https://github.com/wix/ricos/pull/3622) fix polls in json schema
-
 
 ## 8.70.19 (Mar 15, 2022)